+++ src/Paket.Core/DependenciesFile.fs
@@ -15,9 +15,10 @@ type InstallOptions =
     { Strict : bool 
       Redirects : bool 
       ImportTargets : bool
+      CopyLocal : bool
       OmitContent : bool }
 
-    static member Default = { Strict = false; OmitContent = false; Redirects = false; ImportTargets = true}
+    static member Default = { Strict = false; OmitContent = false; Redirects = false; ImportTargets = true; CopyLocal = true}
 
 /// [omit]
 module DependenciesFileParser = 
@@ -134,6 +135,7 @@ module DependenciesFileParser =
     | ReferencesMode of bool
     | OmitContent of bool
     | ImportTargets of bool
+    | CopyLocal of bool
     | Redirects of bool
 
     let private (|Remote|Package|Comment|ParserOptions|SourceFile|) (line:string) =
@@ -164,6 +166,7 @@ module DependenciesFileParser =
         | String.StartsWith "redirects" trimmed -> ParserOptions(ParserOption.Redirects(trimmed.Trim() = "on"))
         | String.StartsWith "content" trimmed -> ParserOptions(ParserOption.OmitContent(trimmed.Trim() = "none"))
         | String.StartsWith "import_targets" trimmed -> ParserOptions(ParserOption.ImportTargets(trimmed.Trim() = "true"))
+        | String.StartsWith "copy_local" trimmed -> ParserOptions(ParserOption.CopyLocal(trimmed.Trim() = "true"))
         | String.StartsWith "gist" _ as trimmed ->
             SourceFile(``parse git source`` trimmed SingleSourceFileOrigin.GistLink "gist")
         | String.StartsWith "github" _ as trimmed  ->
@@ -184,6 +187,7 @@ module DependenciesFileParser =
                 | Comment(line) -> lineNo, options, sources, packages, sourceFiles, if line <> "" then (lineNo,line)::comments else comments
                 | ParserOptions(ParserOption.ReferencesMode mode) -> lineNo, { options with Strict = mode }, sources, packages, sourceFiles, comments
                 | ParserOptions(ParserOption.Redirects mode) -> lineNo, { options with Redirects = mode }, sources, packages, sourceFiles, comments
+                | ParserOptions(ParserOption.CopyLocal mode) -> lineNo, { options with CopyLocal = mode }, sources, packages, sourceFiles, comments
                 | ParserOptions(ParserOption.ImportTargets mode) -> lineNo, { options with ImportTargets = mode }, sources, packages, sourceFiles, comments
                 | ParserOptions(ParserOption.OmitContent omit) -> lineNo, { options with OmitContent = omit }, sources, packages, sourceFiles, comments
                 | Package(name,version,rest) ->
@@ -424,6 +428,7 @@ type DependenciesFile(fileName,options,sources,packages : PackageRequirement lis
             let hasReportedSecond = ref false
             [ if options.Strict then yield "references strict"
               if not options.ImportTargets then yield "import_targets false"
+              if not options.CopyLocal then yield "copy_local false"
               if options.OmitContent then yield "content none"
               for sources, packages in sources do
                   for source in sources do