+++ src/Paket.Core/InstallProcess.fs
@@ -158,8 +158,9 @@ let InstallIntoProjects(sources,force, hard, withBindingRedirects, lockFile:Lock
         let usedPackageSettings =
             usedPackages
             |> Seq.map (fun u -> NormalizedPackageName u.Key,
-                                    { u.Value
-                                        with ImportTargets = u.Value.ImportTargets && lockFile.Options.ImportTargets })
+                                    { u.Value with 
+                                        ImportTargets = u.Value.ImportTargets && lockFile.Options.ImportTargets
+                                        CopyLocal = u.Value.CopyLocal && lockFile.Options.CopyLocal })
             |> Map.ofSeq
 
         project.UpdateReferences(model,usedPackageSettings,hard)