+++ src/Paket.Core/LockFile.fs
@@ -29,6 +29,7 @@ module LockFileSerializer =
               if options.OmitContent then yield "CONTENT: NONE"
               if options.Redirects then yield "REDIRECTS: ON"
               if not options.ImportTargets then yield "IMPORT-TARGETS: FALSE"
+              if not options.CopyLocal then yield "COPY-LOCAL: FALSE"
               for (source, _), packages in sources do
                   if not !hasReported then
                     yield "NUGET"
@@ -119,6 +120,7 @@ module LockFileParser =
     | ReferencesMode of bool
     | OmitContent of bool
     | ImportTargets of bool
+    | CopyLocal of bool
     | Redirects of bool
 
     let private (|Remote|NugetPackage|NugetDependency|SourceFile|RepositoryType|Blank|InstallOption|) (state, line:string) =
@@ -133,6 +135,7 @@ module LockFileParser =
         | _, String.StartsWith "REFERENCES:" trimmed -> InstallOption(ReferencesMode(trimmed.Trim() = "STRICT"))
         | _, String.StartsWith "REDIRECTS:" trimmed -> InstallOption(Redirects(trimmed.Trim() = "ON"))
         | _, String.StartsWith "IMPORT-TARGETS:" trimmed -> InstallOption(ImportTargets(trimmed.Trim() = "TRUE"))
+        | _, String.StartsWith "COPY-LOCAL:" trimmed -> InstallOption(CopyLocal(trimmed.Trim() = "TRUE"))
         | _, String.StartsWith "CONTENT:" trimmed -> InstallOption(OmitContent(trimmed.Trim() = "NONE"))
         | _, trimmed when line.StartsWith "      " ->
             if trimmed.Contains("(") then
@@ -158,6 +161,7 @@ module LockFileParser =
             | InstallOption (ReferencesMode(mode)) -> { state with Options = {state.Options with Strict = mode} }
             | InstallOption (Redirects(mode)) -> { state with Options = {state.Options with Redirects = mode} }
             | InstallOption (ImportTargets(mode)) -> { state with Options = {state.Options with ImportTargets = mode} }
+            | InstallOption (CopyLocal(mode)) -> { state with Options = {state.Options with CopyLocal = mode} }
             | InstallOption (OmitContent(omit)) -> { state with Options = {state.Options with OmitContent = omit} }
             | RepositoryType repoType -> { state with RepositoryType = Some repoType }
             | NugetPackage details ->