+++ RELEASE_NOTES.md
@@ -1,5 +1,8 @@
+#### 0.28.8 - 17.02.2015
+* Allow to opt-out of private references via global switch in `paket.dependencies` - https://github.com/fsprojects/Paket/issues/587
+
 #### 0.28.7 - 17.02.2015
-* Allow to opt-out of targets import via global switch in `paket.dependencies` - https://github.com/fsprojects/Paket/issues/615
+* Allow to opt-out of targets import via global switch in `paket.dependencies` - https://github.com/fsprojects/Paket/issues/587
 
 #### 0.28.6 - 17.02.2015
 * Allow to restrict content files in `paket.references` - https://github.com/fsprojects/Paket/issues/587